login/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
python-dotenv==1.0.0
firebase-admin==6.2.0
PyJWT==2.8.0
python-multipart==0.0.6 